Create event type
GitHub enables integrations to subscribe to events that happen on the
platform. Whenever an event happens, a webhook is sent to the
integration with the event as its payload.

A new `enum` has been created to represent these events. It has a
catch-all variant for events that the crate does not support yet, which
wraps the webhook payload as a JSON value.