Create authentication tokens for GitHub
A new module has been added with structs for an app and installation
token, which are used to authenticate with GitHub's API. The app token
represents the GitHub App itself, while the installation token
represents the installation of the app in a specific account. It can be
used to access resources that belongs to the owner of the account.