Create action to get a file from GitHub (#19)
A new action has been prototyped that can get a file from a repository
on GitHub. When the action is succesfull, it returns a file object with
the decoded file content. Otherwise, an error is returned.

The action is very much a prototype, and its interface and return types
will very likely change as we start using it.