Handle not found error when getting a file (#20)
When a user tries to get a file from GitHub that doesn't exist, GitHub
responds with a HTTP 404 NOT FOUND response. The response was previously
not caught, and caused the workflow to fail completely. This has been
fixed, and an error response from GitHub is now converted to a
meaningful error inside the crate.