Refactor error types
The error types in the crate have been refactored to increase their
usability. Overly verbose variants of both error types have been
removed, and the anyhow crate has been added to capture context around
unexpected errors.

This is just the first step on a long journey to provide a clean API for
errors that can occur within the crate. As we gain more experience with
actions and the workflows built on top of them, we will most likely
refactor the errors at least once.