Remove documentation
When starting the crate, we enabled a lint that enforced documentation
for public interfaces. This was premature, since we are still learning a
lot and the API of the crate is changing as a consequence. As a result,
most comments where not helpful and just placeholders.

The lint has been removed, as have all existing comments. When the crate
has stabilized, we will reintroduce the lint and write proper
documentation then.