Create trait for actions (#31)
Actions all follow the same pattern: They accept an input, interact with
GitHub, and then return a result. As we are creating more actions, we
want to capture as much shared logic as possible and make it reusable.
For this purpose, a new trait has been created that describes an action.

The action to list check runs has been refactored already. The action to
get a file however has a slightly different internal logic that is not
(yet) compatible with the trait.