Add HEAD SHA to check run entity (#34)
The check run webhook from GitHub contains the SHA of the commit for
which the check run was triggered. When an app wants to create a check
run for the same commit, it must use the SHA to do so. It has therefore
been added to the check run entity as a field.