Add getter for login to account entity
The account entity is an enumeration with an organization and a user
variant. While the fields on both can slightly differ, they both have a
`login` as their identifier. A new convenience method has been added to
the account entity to return the login for both variants. Consumers
don't need to match the variant anymore.