Cache token in GitHub client
The GitHub client has been refactored to cache its tokens as long as
they are valid. Once a token expires, it is recreated and cached again.
This reduces the number of outgoing requests to the GitHub API by half,
since each request always requested its own installation token.

The client now maintains an internal token factory, which required
mutability to cache the tokens. This also required a change to the
`Action` trait, which might turn out to be either good or bad. We
consider this only the first of many changes to the public interface of
both the client and the action trait, which will change as we learn more
about the domain and their usage.