Pass GitHub Client by reference
The GitHub client is passed by reference again, allowing different steps
in a workflow to reuse the same client and thus token. While this makes
it easy to reuse the client and token, it is not a very ergonomic
solution. We might refactor this again and use a mutex or similar to
achieve internal mutability in the token factory.