Change GitHub client to interior mutability (#53)
The GitHub client and the token factory have been refactored to use
interior mutability. This makes it easier to pass shared references
around, for example when building a workflow in octox.

Because every call into the factory now requires at least one mutex to
be locked, the solution is less ideal than the previous one. But the
ergonomics are greatly improved, and even this solution is better than
the original implementation that sent a request to GitHub's API for
every token.